+++ src/components/Pages/DashboardPage/DashboardPage.module.scss
@@ -2,7 +2,7 @@
   min-height: 100vh;
   width: 100%;
 
-  backdrop-filter: blur(25px);
+  // backdrop-filter: blur(25px);
 }
 .gridPage {
   display: grid;
