+++ src/components/Pages/TransactionsPage/TransactionsPage.jsx
@@ -1,16 +1,27 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 // import PropTypes from 'prop-types';
 import TableMobile from 'components/TableMobile/TableMobile';
 import TableTransactions from 'components/TableTransactions/TableTransactions';
-
-import Table from '../NotFoundPage/Table/Table';
+import TableContext from 'components/Table/TableContext';
+import Table from 'components/Table/Table';
 import {
   NotMobile,
   Mobile,
 } from 'components/DeviceTypeControl/DeviseTypeController';
+
+import { useDispatch, useSelector } from 'react-redux';
+import {
+  getTransactionsThunk,
+  addTransactionThunk,
+} from 'redux/finance/finance-operation';
 import scss from './TransactionsPage.module.scss';
 
 const TransactionsPage = props => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(getTransactionsThunk());
+  }, [dispatch]);
   return (
     <div className={scss.TransactionsPage}>
       <Mobile>
@@ -25,7 +36,9 @@ const TransactionsPage = props => {
             illum. Maiores enim asperiores dolore saepe aut!
           </span> */}
           {/* <TableTransactions /> */}
-          <Table />
+          <TableContext>
+            <Table tableData></Table>
+          </TableContext>
         </div>
       </NotMobile>
     </div>
