+++ src/components/Table/Collumns/Coll_Sum.jsx
