+++ src/components/Table/Collumns/Coll_Type.jsx
