+++ src/components/Table/Table.module.scss
@@ -47,20 +47,22 @@
 .isRowOpen {
   transform: scale(1.05);
   z-index: 500;
+
   transition: all 250ms cubic-bezier(0.4, 0, 0.2, 1);
   &:hover {
     box-shadow: 0 0 5px rgba(126, 126, 126, 0.5);
   }
   & .rowTop {
-    max-height: fit-content;
+    z-index: 500;
     border-radius: 30px 30px 0 0;
-    // backdrop-filter: blur(25px);
+    backdrop-filter: blur(15px);
 
   }
   & .rowBottom {
+    z-index: 500;
     max-height: fit-content;
     border-radius: 0 0 30px 30px;
-    // backdrop-filter: blur(25px);
+    backdrop-filter: blur(15px);
   }
 }
 .rowTop {
@@ -91,7 +93,6 @@
 
   max-height: 0px;
 
-  backdrop-filter: blur(25px);
   transition: all 250ms cubic-bezier(0.4, 0, 0.2, 1);
 }
 // * ============= FORM
